rm-rfp.rs: Add --interactive mode
It's not quite like normal `rm -ri` because it's asking in the deletion part
and not in the finding part. The though was the user could take their time
and the finder could continue collecting files to delete, but the downside
is it prevents the user from being able to prevent descending into a
directory.