rm-rfp.rs, integration.rs: Add POSIX safety measures and test for them
Adds the --no-preserve-root option. The default is to fail if one of the
command line arguments is / or something that is the equivalent
(inode-wise).

This also adds a check for . and .. in the last position, which POSIX
apparently forbids.

This also adds a check equivalent to GNU coreutil's
--preserve-root=all. This check defaults to on and is also disabled by the
--no-preserve-root option.

I _hate_ the {no-,}preserve-root tests. It tries very hard to be safe but it
really shouldn't even be toying with things the way it does. Unfortunately I
don't know how to set up a chroot or sandbox with only user permissions
(needing to be root to run tests is even worse).